+++ 2508date/src/NoC/VCRouter.hpp
@@ -20,10 +20,10 @@ class VCNetwork;
 extern unsigned int cycles;
 extern  std::vector<std::vector<int>> yzLeaveOutportPerRouter;
 extern  std::vector<std::vector<int>> yzEnterInportPerRouter;
-extern int  YZGlobalFlitPass;  // Total hop count (router + NI)
-extern int  YZGlobalRouterHopCount;  // Router-only hop count
-extern int  YZGlobalNIHopCount;  // NI-only hop count
-extern int   YZGlobalRespFlitPass;
+extern long long  YZGlobalFlitPass;  // Total hop count (router + NI)
+extern long long YZGlobalRouterHopCount;  // Router-only hop count
+extern long long YZGlobalNIHopCount;  // NI-only hop count
+extern long long   YZGlobalRespFlitPass;
 
 class VCRouter: public NRBase
 {
